+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## Unreleased
 ### Added
+- Console command for reindexing to Elasticsearch.
 - Example page transformers and short description in README.
 
 ## 0.2.3 - 2017-09-07